move FullcreenControl example to examples/controls
