+++ examples/custom-controller/app.css
@@ -31,3 +31,7 @@ label {
 input {
   margin-left: 20px;
 }
+
+.control-group {
+  display: flex;
+}
