+++ examples/draggable-markers/src/app.js
@@ -1,7 +1,7 @@
 /* global window */
 import React, {Component} from 'react';
 import {render} from 'react-dom';
-import MapGL, {Marker, Popup, NavigationControl} from 'react-map-gl';
+import MapGL, {Marker, NavigationControl, FullscreenControl} from 'react-map-gl';
 
 import ControlPanel from './control-panel';
 import Pin from './pin';
