--- examples/fullscreen-map/README.md
@@ -1,12 +1,0 @@
-<div align="center">
-  <img src="https://avatars3.githubusercontent.com/u/2105791?v=3&s=200" />
-</div>
-
-## Example: Fullscreen Map 
-
-Demonstrates how to toggle between current view and fullscreen mode.
-
-```
-    npm install
-    npm start
-```
