--- examples/fullscreen-map/app.css
@@ -1,8 +1,0 @@
-body {
-  margin: 0;
-  background: #000;
-}
-#map {
-  width: 100vw;
-  height: 100vh;
-}
