--- examples/fullscreen-map/index.html
@@ -1,15 +1,0 @@
-<!doctype html>
-<html>
-<head>
-  <meta charset='UTF-8' />
-  <title>react-map-gl Example</title>
-  <link rel="stylesheet" type="text/css" href="app.css" />
-</head>
-<body>
-<div id="map"></div>
-<script src='app.js'></script>
-</body>
-<script type="text/javascript">
-  App.renderToDom(document.getElementById('map'));
-</script>
-</html>
